requirements.txt
amqp==5.3.1
billiard==4.2.4
blinker==1.9.0
camelot-py==1.0.9
celery==5.6.0
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
configparser==7.2.0
cryptography==46.0.3
et_xmlfile==2.0.0
exceptiongroup==1.3.1
Flask==3.1.2
gitdb==4.0.12
GitPython==3.1.41
greenlet==3.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.6.1
MarkupSafe==3.0.3
numpy==2.2.6
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pdfminer.six==20251107
pillow==12.0.0
prompt_toolkit==3.0.52
pycparser==2.23
pypdf==5.9.0
pypdfium2==5.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
setuptools==78.1.1
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.45
tabulate==0.9.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.4
